Merged leveldb and leveldb-lin
The only difference between leveldb and leveldb-lin was that
leveldb-lin has a Makefile that is absent from leveldb directory.

As Makefile does not conflict with original leveldb directory
at any way - moved it from leveldb-lin to leveldb and delted
the leveldb-lin direcotry completly.

Now the process of compiling the leveldb for all platforms
is exactly the same with no need to move/rename things around